requirements.txt
flask
youtube-transcript-api
langchain
langchain-community
langchain-huggingface
langchain-text-splitters
faiss-cpu
python-dotenv